refactor: remove support for very old scully versions
BREAKING CHANGE:
- drop support for scully versions <0.0.85

closes #28
