+++ package.json
@@ -4,13 +4,21 @@
   "description": "GitHub Action to build and deploy your Scully site to GitHub Pages ❤️🎩",
   "main": "index.js",
   "scripts": {
-    "build": "tsc --noEmit && ncc build index.ts -o dist -m",
-    "test": "tsc --noEmit && jest --silent",
-    "lint": "eslint . --ext .ts",
-    "prepare": "husky install"
+    "bundle": "npm run format:write && npm run package",
+    "ci-test": "jest",
+    "format:write": "prettier --write **/*.ts",
+    "format:check": "prettier --check **/*.ts",
+    "lint": "npx eslint . -c ./.github/linters/.eslintrc.yml",
+    "package": "ncc build src/index.ts --license licenses.txt",
+    "package:watch": "npm run package -- --watch",
+    "test": "jest",
+    "all": "npm run format:write && npm run lint && npm run test && npm run package"
+  },
+  "exports": {
+    ".": "./dist/index.js"
   },
   "engines": {
-    "node": "^18 || ^20"
+    "node": ">=20"
   },
   "repository": {
     "type": "git",
@@ -27,29 +35,58 @@
     "url": "https://github.com/cmgriffing/scully-gh-pages-action/issues"
   },
   "homepage": "https://github.com/cmgriffing/scully-gh-pages-action#readme",
+  "jest": {
+    "preset": "ts-jest",
+    "verbose": true,
+    "clearMocks": true,
+    "testEnvironment": "node",
+    "moduleFileExtensions": [
+      "js",
+      "ts"
+    ],
+    "testMatch": [
+      "**/*.test.ts"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/"
+    ],
+    "transform": {
+      "^.+\\.ts$": "ts-jest"
+    },
+    "coverageReporters": [
+      "json-summary",
+      "text",
+      "lcov"
+    ],
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "./src/**"
+    ]
+  },
   "dependencies": {
-    "@actions/core": "^1.6.0",
-    "@actions/exec": "^1.1.0",
-    "@actions/github": "^5.0.0",
-    "@actions/io": "^1.1.1",
-    "@yarnpkg/lockfile": "1.1.0",
-    "semver": "^7.5.2"
+    "@actions/core": "^1.10.1",
+    "@actions/exec": "^1.1.1",
+    "@actions/github": "^5.1.1",
+    "@actions/io": "^1.1.3"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.3",
-    "@types/node": "^20.10.0",
-    "@types/semver": "^7.3.9",
-    "@types/yarnpkg__lockfile": "^1.1.5",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "@typescript-eslint/parser": "^5.7.0",
-    "@zeit/ncc": "^0.22.3",
-    "eslint": "^8.4.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^7.0.0",
-    "jest": "^27.4.4",
-    "prettier": "^2.5.1",
-    "ts-jest": "^27.1.1",
-    "typescript": "^4.5.3"
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.9.2",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "@vercel/ncc": "^0.38.1",
+    "eslint": "^8.54.0",
+    "eslint-plugin-github": "^4.10.1",
+    "eslint-plugin-jest": "^27.6.0",
+    "eslint-plugin-jsonc": "^2.10.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "husky": "^8.0.3",
+    "jest": "^29.7.0",
+    "js-yaml": "^4.1.0",
+    "prettier": "^3.1.0",
+    "prettier-eslint": "^16.1.2",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.2.2"
   }
 }
